workload handler layer

+++ profileService/internal/handler/facultyProfile/model.go
@@ -152,11 +152,11 @@ type WorkloadStats struct {
 }
 
 type Classes struct {
-	Lec  int `json:"lec_hours"`
-	Tut  int `json:"tut_hours"`
-	Lab  int `json:"lab_hours"`
-	Elec int `json:"elective_hours"`
-	Rate int `json:"rate"`
+	Lec  int     `json:"lec_hours"`
+	Tut  int     `json:"tut_hours"`
+	Lab  int     `json:"lab_hours"`
+	Elec int     `json:"elective_hours"`
+	Rate float64 `json:"rate"`
 }
 
 type GetAllFacultiesResponse struct {
